✨ feat: The total is auto-calculated in Order and Order Item

+++ src/main/java/br/fatec/easycoast/dtos/product/ProductRequest.java
@@ -17,9 +17,9 @@ public record ProductRequest(
 
                 @NotBlank(message = "Description cannot be blank") @Size(min = 10, message = "Minimum length for the description is 10 characters") String description,
 
-                @NotNull(message = "Price cannot be null") @Positive(message = "Price must be positive") float price,
+                @NotNull(message = "Price cannot be null") @Positive(message = "Price must be positive") double price,
 
-                @NotNull(message = "Discount cannot be null") @Min(value = 0, message = "Discount must be at least 0") @Max(value = 100, message = "Discount cannot be more than 100") float discount,
+                @NotNull(message = "Discount cannot be null") @Min(value = 0, message = "Discount must be at least 0") @Max(value = 100, message = "Discount cannot be more than 100") double discount,
 
                 @NotNull(message = "Availability cannot be null") Boolean availability,
